Refactor logic into forecast.py
